+++ nf_core/modules/install.py
@@ -3,7 +3,8 @@ import os
 
 import questionary
 
-import nf_core.modules.module_utils
+import nf_core.components.components_install
+import nf_core.modules.modules_utils
 import nf_core.utils
 from nf_core.modules.modules_json import ModulesJson
 
@@ -44,36 +45,18 @@ class ModuleInstall(ModuleCommand):
         modules_json = ModulesJson(self.dir)
         modules_json.check_up_to_date()
 
-        if self.prompt and self.sha is not None:
-            log.error("Cannot use '--sha' and '--prompt' at the same time!")
+        # Verify SHA
+        if not self.modules_repo.verify_sha(self.prompt, self.sha):
             return False
 
-        # Verify that the provided SHA exists in the repo
-        if self.sha:
-            if not self.modules_repo.sha_exists_on_branch(self.sha):
-                log.error(f"Commit SHA '{self.sha}' doesn't exist in '{self.modules_repo.remote_url}'")
-                return False
-
-        if module is None:
-            module = questionary.autocomplete(
-                "Tool name:",
-                choices=self.modules_repo.get_avail_modules(),
-                style=nf_core.utils.nfcore_question_style,
-            ).unsafe_ask()
-
-        # Check that the supplied name is an available module
-        if module and module not in self.modules_repo.get_avail_modules():
-            log.error(f"Module '{module}' not found in list of available modules.")
-            log.info("Use the command 'nf-core modules list' to view available software")
+        # Check and verify module name
+        module = nf_core.components.components_install.collect_and_verify_name(
+            self.component_type, module, self.modules_repo
+        )
+        if not module:
             return False
 
-        if not self.modules_repo.module_exists(module):
-            warn_msg = (
-                f"Module '{module}' not found in remote '{self.modules_repo.remote_url}' ({self.modules_repo.branch})"
-            )
-            log.warning(warn_msg)
-            return False
-
+        # Get current version
         current_version = modules_json.get_module_version(
             module, self.modules_repo.remote_url, self.modules_repo.repo_path
         )
@@ -85,61 +68,30 @@ class ModuleInstall(ModuleCommand):
         module_dir = os.path.join(install_folder, module)
 
         # Check that the module is not already installed
-        if (current_version is not None and os.path.exists(module_dir)) and not self.force:
-            log.info("Module is already installed.")
+        if not nf_core.components.components_install.check_component_installed(
+            self.component_type, module, current_version, module_dir, self.modules_repo, self.force
+        ):
+            return False
 
-            self.force = questionary.confirm(
-                f"Module {module} is already installed. Do you want to force the reinstallation?",
-                style=nf_core.utils.nfcore_question_style,
-                default=False,
-            ).unsafe_ask()
+        version = nf_core.components.components_install.get_version(
+            module, self.component_type, self.sha, self.prompt, current_version, self.modules_repo
+        )
+        if not version:
+            return False
 
-            if not self.force:
-                repo_flag = (
-                    "" if self.modules_repo.repo_path == NF_CORE_MODULES_NAME else f"-g {self.modules_repo.remote_url} "
-                )
-                branch_flag = "" if self.modules_repo.branch == "master" else f"-b {self.modules_repo.branch} "
-
-                log.info(
-                    f"To update '{module}' run 'nf-core modules {repo_flag}{branch_flag}update {module}'. To force reinstallation use '--force'"
-                )
-                return False
-
-        if self.sha:
-            version = self.sha
-        elif self.prompt:
-            try:
-                version = nf_core.modules.module_utils.prompt_module_version_sha(
-                    module,
-                    installed_sha=current_version,
-                    modules_repo=self.modules_repo,
-                )
-            except SystemError as e:
-                log.error(e)
-                return False
-        else:
-            # Fetch the latest commit for the module
-            version = self.modules_repo.get_latest_module_version(module)
-
+        # Remove module if force is set
         if self.force:
             log.info(f"Removing installed version of '{self.modules_repo.repo_path}/{module}'")
-            self.clear_module_dir(module, module_dir)
-            for repo_url, repo_content in modules_json.modules_json["repos"].items():
-                for dir, dir_modules in repo_content["modules"].items():
-                    for name, _ in dir_modules.items():
-                        if name == module and dir == self.modules_repo.repo_path:
-                            repo_to_remove = repo_url
-                            log.info(
-                                f"Removing module '{self.modules_repo.repo_path}/{module}' from repo '{repo_to_remove}' from modules.json"
-                            )
-                            modules_json.remove_entry(module, repo_to_remove, self.modules_repo.repo_path)
-                            break
+            self.clear_component_dir(module, module_dir)
+            nf_core.components.components_install.clean_modules_json(
+                module, self.component_type, self.modules_repo, modules_json
+            )
 
         log.info(f"{'Rei' if self.force else 'I'}nstalling '{module}'")
         log.debug(f"Installing module '{module}' at modules hash {version} from {self.modules_repo.remote_url}")
 
         # Download module files
-        if not self.install_module_files(module, version, self.modules_repo, install_folder):
+        if not self.install_component_files(module, version, self.modules_repo, install_folder):
             return False
 
         if not silent: