+++ nf_core/subworkflows/__init__.py
@@ -2,4 +2,5 @@ from .create import SubworkflowCreate
 from .info import SubworkflowInfo
 from .install import SubworkflowInstall
 from .list import SubworkflowList
+from .subworkflows_test import SubworkflowsTest
 from .test_yml_builder import SubworkflowTestYmlBuilder